add ont_lock_proxy contract
